Broke up jupyter.py methods into a sub-package for all jupyter utilities. Added some additional code comments
